+++ extensions/spark/kyuubi-spark-authz/pom.xml
@@ -23,7 +23,7 @@
         <artifactId>kyuubi-parent</artifactId>
         <groupId>org.apache.kyuubi</groupId>
         <version>1.6.0-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../../../pom.xml</relativePath>
 
     </parent>
     <modelVersion>4.0.0</modelVersion>
@@ -229,6 +229,12 @@
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>jul-to-slf4j</artifactId>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
