+++ extensions/spark/kyuubi-spark-authz/src/test/resources/log4j2-test.properties
