+++ extensions/spark/kyuubi-spark-authz/src/test/resources/sparkSql_hive_jenkins.json
